+++ examples/address-validation.php
@@ -1,71 +1,38 @@
 <?php
-/**
- * This test will send the same test data as in FedEx's documentation:
- * /php/RateAvailableServices/RateAvailableServices.php5
- */
 
-//remember to copy example.credentials.php as credentials.php replace 'FEDEX_KEY', 'FEDEX_PASSWORD', 'FEDEX_ACCOUNT_NUMBER', and 'FEDEX_METER_NUMBER'
 require_once 'credentials.php';
 require_once 'bootstrap.php';
 
-use FedEx\AddressValidationService;
+use FedEx\AddressValidationService\Request;
 use FedEx\AddressValidationService\ComplexType;
 use FedEx\AddressValidationService\SimpleType;
 
-$userCredential = new ComplexType\WebAuthenticationCredential();
-$userCredential
-    ->setKey(FEDEX_KEY)
-    ->setPassword(FEDEX_PASSWORD);
-
-$webAuthenticationDetail = new ComplexType\WebAuthenticationDetail();
-$webAuthenticationDetail->setUserCredential($userCredential);
-
-$clientDetail = new ComplexType\ClientDetail();
-$clientDetail
-    ->setAccountNumber(FEDEX_ACCOUNT_NUMBER)
-    ->setMeterNumber(FEDEX_METER_NUMBER);
-
-$version = new ComplexType\VersionId();
-$version
-    ->setMajor(2)
-    ->setIntermediate(0)
-    ->setMinor(0)
-    ->setServiceId('aval');
-
-$options = new ComplexType\AddressValidationOptions();
-$options
-    ->setCheckResidentialStatus(true)
-    ->setVerifyAddresses(true)
-    ->setConvertToUpperCase(true)
-    ->setDirectionalAccuracy(SimpleType\AddressValidationAccuracyType::_MEDIUM);
-
-$addresses = array();
-
-$address1 = new ComplexType\Address();
-$address1
-    ->setStreetLines(array('12345 Main Street'))
-    ->setCity('Anytown')
-    ->setStateOrProvinceCode('NY')
-    ->setPostalCode('47711')
-    ->setCountryCode('US');
-
-$addressToValidate1 = new ComplexType\AddressToValidate();
-$addressToValidate1
-    ->setAddress($address1)
-    ->setAddressId(1);
-
-$addresses[] = $addressToValidate1;
-
 $addressValidationRequest = new ComplexType\AddressValidationRequest();
-$addressValidationRequest->setWebAuthenticationDetail($webAuthenticationDetail);
-$addressValidationRequest->setClientDetail($clientDetail);
-$addressValidationRequest->setVersion($version);
-$addressValidationRequest->setRequestTimestamp(date('c'));
-$addressValidationRequest->setOptions($options);
-$addressValidationRequest->setAddressesToValidate($addresses);
 
+// User Credentials
+$addressValidationRequest->WebAuthenticationDetail->UserCredential->Key = FEDEX_KEY;
+$addressValidationRequest->WebAuthenticationDetail->UserCredential->Password = FEDEX_PASSWORD;
 
-$validateShipmentRequest = new AddressValidationService\Request();
-$result = $validateShipmentRequest->getAddressValidationReply($addressValidationRequest);
+// Client Detail
+$addressValidationRequest->ClientDetail->AccountNumber = FEDEX_ACCOUNT_NUMBER;
+$addressValidationRequest->ClientDetail->MeterNumber = FEDEX_METER_NUMBER;
 
-var_dump($result);
+// Version
+$addressValidationRequest->Version->ServiceId = 'aval';
+$addressValidationRequest->Version->Major = 4;
+$addressValidationRequest->Version->Intermediate = 0;
+$addressValidationRequest->Version->Minor = 0;
+
+// Address(es) to validate.
+$addressValidationRequest->AddressesToValidate = [new ComplexType\AddressToValidate()]; // just validating 1 address in this example.
+$addressValidationRequest->AddressesToValidate[0]->Address->StreetLines = ['12345 Main Street'];
+$addressValidationRequest->AddressesToValidate[0]->Address->City = 'Anytown';
+$addressValidationRequest->AddressesToValidate[0]->Address->StateOrProvinceCode = 'NY';
+$addressValidationRequest->AddressesToValidate[0]->Address->PostalCode = 47711;
+$addressValidationRequest->AddressesToValidate[0]->Address->CountryCode = 'US';
+
+$request = new Request();
+$addressValidationReply = $request->getAddressValidationReply($addressValidationRequest, true);
+
+var_dump($addressValidationReply);
+