+++ src/FedEx/AddressValidationService/ComplexType/WebAuthenticationCredential.php
@@ -4,7 +4,7 @@ namespace FedEx\AddressValidationService\ComplexType;
 use FedEx\AbstractComplexType;
 
 /**
- * Two part authentication string used for the sender's identity.
+ * Two part authentication string used for the sender's identity
  *
  * @author      Jeremy Dunn <jeremy@jsdunn.info>
  * @package     PHP FedEx API wrapper
@@ -24,7 +24,7 @@ class WebAuthenticationCredential extends AbstractComplexType
     protected $name = 'WebAuthenticationCredential';
 
     /**
-     * Identifying part of the authentication key. This value is provided by FedEx after registration.
+     * Identifying part of authentication credential. This value is provided by FedEx after registration
      *
      * @param string $key
      * @return $this
@@ -36,7 +36,7 @@ class WebAuthenticationCredential extends AbstractComplexType
     }
 
     /**
-     * Secret part of authentication key used for authentication. This value is provided by FedEx after registration.
+     * Secret part of authentication key. This value is provided by FedEx after registration.
      *
      * @param string $password
      * @return $this