+++ src/FedEx/AddressValidationService/SimpleType/FedExAddressClassificationType.php
@@ -4,18 +4,16 @@ namespace FedEx\AddressValidationService\SimpleType;
 use FedEx\AbstractSimpleType;
 
 /**
- * ResidentialStatusType
+ * Specifies the address classification (business vs. residential)
  *
  * @author      Jeremy Dunn <jeremy@jsdunn.info>
  * @package     PHP FedEx API wrapper
  * @subpackage  Address Validation Service
  */
-class ResidentialStatusType extends AbstractSimpleType
+class FedExAddressClassificationType extends AbstractSimpleType
 {
-    const _UNDETERMINED = 'UNDETERMINED';
     const _BUSINESS = 'BUSINESS';
+    const _MIXED = 'MIXED';
     const _RESIDENTIAL = 'RESIDENTIAL';
-    const _INSUFFICIENT_DATA = 'INSUFFICIENT_DATA';
-    const _UNAVAILABLE = 'UNAVAILABLE';
-    const _NOT_APPLICABLE_TO_COUNTRY = 'NOT_APPLICABLE_TO_COUNTRY';
+    const _UNKNOWN = 'UNKNOWN';
 }