+++ util/Cli/GenerateCode/Command/GenerateCode.php
@@ -114,7 +114,7 @@ class GenerateCode extends Console\Command\Command
 
 
         //AddressValidationService
-        $wsdlPath = $fedexSrcDir . '/_wsdl/AddressValidationService_v2.wsdl';
+        $wsdlPath = $fedexSrcDir . '/_wsdl/AddressValidationService_v4.wsdl';
 
         $baseNamespace = 'FedEx\AddressValidationService';
         $subPackageName = 'Address Validation Service';